refactor : validateInput 폴더 위치 utils 하위로 변경
